backend/foodgram/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.19
django-colorfield==0.6.3
django-debug-toolbar==3.2.1
django-extra-fields==3.0.2
django-filter==21.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sorl-thumbnail==12.7.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0